fix: add a heartbeat message to Duo Workflow bidi stream
We have discovered that our infrastructure will not keep our long
running Duo Workflow gRPC streams alive beyond 100s of inactivity. There
are lower level protocols (keepalive pings) that we've already put in
place but through our investigation we have determined these are not
sufficient and Cloudflare will always disconnect us if there aren't any
application messages sent every 100s. As such we are adding a heartbeat
message type to the stream.

Initially we just need the Duo Workflow Service to ignore these messages
and then we can do tests from the Executor to confirm that this does
indeed fix our problem. If it turns out that this doesn't work OR if we
can find another way to achieve this we may end up reverting this before
we actually start sending this from the executor.

Read more at
https://gitlab.com/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/issues/1274
and
https://gitlab.com/gitlab-org/gitlab/-/issues/509586 .