refactor: unify AIGW and DWS metrics
As a first step towards
https://gitlab.com/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/issues/1626,
introduce the labels that DWS expects to the shared metrics. Notably though,
we don't treat the `stop_reason` as the determinant for success, but still include it
in the metrics dimensions (see
https://gitlab.com/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/issues/1626#note_2861839934).
The DWS counter and time metrics are redundant now, but are maintained to facilitate the
transition. After the rollout of the Runbooks changes, we can delete the DWS-specific counters.